requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
websockets
rich
python-dotenv
